UserDiscountRepository created, UserRepo modified

+++ src/main/java/repository/UserRepository.java
@@ -1,12 +1,14 @@
 package repository;
 
 
+import java.math.BigDecimal;
+
 import model.user.User;
 
 public interface UserRepository extends Repository<User> {
 	
 	public User findByEmail(String email);
 	
-	public long getSpentMoney(long user_id);
+	public BigDecimal getSpentMoney(long user_id);
 	
 }
